requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
django-filter==22.1
Pillow==9.4.0
sqlparse==0.4.3
psycopg2==2.8.4
numpy~=1.21.6
